1.0 version of parser completed

+++ src/main/java/org/uka0001/Person.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 public class Person {
     private String name;
     private int age;
-    private float salary;
+    private int salary;
     private String position;
     private LocalDate dateOfBirth;
